@@ -1,9 +1,8 @@
-module.exports = {
-  content: [
-    './src/**/*.{js,ts,jsx,tsx}', // Adjust the paths according to your project structure
-  ],
+/** @type {import('tailwindcss').Config} */
+export default {
+  content: ["./src/**/*.{js,ts,jsx,tsx}"],
   theme: {
     extend: {},
   },
   plugins: [],
-}
+};
